GPIO: - Interrupts now call a handler object instead of a function pointer
TSTP:
- Changed Hash tables for simple lists
- Changed times to 64bits
- Accesses to data are now volatile
- Began wiegand code cleanup


git-svn-id: https://epos.lisha.ufsc.br/svn/epos2@3881 86035c5f-d46c-4136-a2d5-fe63f1d4a782

+++ branches/arm/include/tstp_time.h
@@ -22,7 +22,7 @@ class PTS
     bool bootstrap();
 
 public:
-    typedef int Time;
+    typedef long long Time;
 
     void frame_reception() { _t1 = time_now(); }
 
@@ -45,7 +45,7 @@ private:
     }
 
     void update_interest(char * header);
-    void update_data(char * header) { 
+    void update_data(char * header) {
         if(not Traits<TSTP>::is_sink) {
             do_update_data(header); 
         }
